fix: handle dark/light mode

+++ src/pages/Lists.jsx
@@ -10,9 +10,11 @@ import { fetchLists, removeList } from 'redux/list/listSlice';
 const Lists = ({ columns }) => {
   const dispatch = useDispatch();
   const { lists } = useSelector((state) => state.lists);
+  const { darkMode } = useSelector((state) => state.darkMode);
 
   const handleDelete = (id) => {
-    dispatch(removeList(id));
+    if (window.confirm('Are you sure you wanted to delete this list'))
+      dispatch(removeList(id));
   };
 
   useEffect(() => {
@@ -36,11 +38,8 @@ const Lists = ({ columns }) => {
             </Link>
             <DeleteOutline
               onClick={() => handleDelete(params.row._id)}
-              style={{
-                fontSize: '2rem',
-                color: '#ff0000',
-                cursor: 'pointer',
-              }}
+              style={{ fontSize: '2rem', cursor: 'pointer' }}
+              className={darkMode ? 'delete__dark' : 'delete__light'}
             />
           </>
         );
@@ -59,6 +58,7 @@ const Lists = ({ columns }) => {
         rowsPerPageOptions={[8]}
         checkboxSelection
         style={{ fontSize: '1.5rem' }}
+        className={darkMode && 'data__grid'}
       />
     </Container>
   );
@@ -67,6 +67,7 @@ const Lists = ({ columns }) => {
 const Container = styled.div`
   flex: 4;
   padding: 2rem;
+  background-color: ${({ theme }) => theme.bg};
 `;
 
 const EditButton = styled.button`
@@ -74,8 +75,8 @@ const EditButton = styled.button`
   display: block;
   padding: 0.5rem 1rem;
   text-transform: capitalize;
-  background-color: #3bb077;
-  color: var(--color-white);
+  background-color: ${({ theme }) => theme.btnEdit};
+  color: ${({ theme }) => theme.textEdit};
   border-radius: 10rem;
   cursor: pointer;
   margin-right: 1rem;
